requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==5.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
azure-core==1.25.1
azure-iot-hub==2.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
ci-info==0.3.0
classification-models-3D==1.0.6
click==8.1.3
colorama==0.4.5
configobj==5.0.8
configparser==7.0.0
cryptography==38.0.1
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.1
docutils==0.19
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.1.3
executing==1.0.0
fastjsonschema==2.16.1
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.2
fonttools==4.37.1
fpdf==1.7.2
frozenlist==1.3.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==7.0.1
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter-server-proxy==3.2.2
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.4.4
looseversion==1.3.0
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
msrest==0.7.1
multidict==6.0.2
mypy-extensions==1.0.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
notebook==6.4.12
numpy==1.23.2
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pbkdf2==1.3
pdf2image==1.17.0
pickleshare==0.7.5
pikepdf==8.13.0
pillow==10.2.0
platformdirs==4.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.22.1
prov==2.0.1
psutil==5.9.2
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
pydot==2.0.0
pyglet==1.5.26
Pygments==2.13.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyOpenGL==3.1.6
pyparsing==3.0.9
PyPDF2==3.0.1
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-time==0.3.0
python-version==0.0.2
python-wifi==0.6.1
python-youtube==0.9.1
pytube==15.0.0
pytube3==9.6.4
pytz==2024.1
pywifi==1.1.12
pywin32==304
pywinpty==2.0.7
pyxnat==1.6.2
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
random2==1.0.1
rdflib==6.3.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scipy==1.13.1
selenium==4.12.0
Send2Trash==1.8.0
simpervisor==0.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socket-client-server==0.5.dev3
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.5.0
streamlit==1.31.1
subprocess.run==0.0.8
tenacity==8.2.3
terminado==0.15.0
terminaltables==3.1.0
tiny-3d-engine==0.2.1
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.1
tornado==6.2
traitlets==5.3.0
traits==6.3.2
trio==0.22.2
trio-websocket==0.10.4
turtles==1.0.0
txaio==22.2.1
typing-inspect==0.9.0
typing_extensions==4.3.0
tzdata==2023.4
tzlocal==5.2
uamqp==1.6.0
uritemplate==4.1.1
urllib3==1.26.12
validators==0.22.0
virtualenv==20.25.0
vpython==7.6.4
watchdog==4.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==3.0.1
widgetsnbextension==4.0.3
wifi==0.3.8
wifi-conf==0.1.dev0
wifi-scan==0.0.1
wireless==0.3.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.8.1
youtube-dl==2021.12.17
zipp==3.17.0
